feat: add Questão 3 - Comparando notas

+++ src/main/java/org/example/Main.java
@@ -1,15 +1,15 @@
 package org.example;
 import java.util.Scanner;
 
-// Questão 2 - Comparação de notas
+// Questão 3 - Comparação de notas
 public class Main {
     public static void main(String[] args) {
         Scanner entrada = new Scanner(System.in);
-        int n1, n2;
+        double n1, n2;
         System.out.println("Digite o primeiro número: ");
-        n1 = entrada.nextInt();
+        n1 = entrada.nextDouble();
         System.out.println("Digite o segundo número: ");
-        n2 = entrada.nextInt();
+        n2 = entrada.nextDouble();
 
         if (n1 > n2 ){
             System.out.println(n1);
